chore: move JKS functions to dedicated file
